Split module and worker again to use separate jsconfig
