+++ module/r2/types.d.ts
@@ -1,3 +1,5 @@
+import type { R2Object, R2StringChecksums, R2HTTPMetadata, R2MultipartUpload } from "@cloudflare/workers-types/experimental";
+
 export type Dispatch = (
   operation: string,
   parameters: unknown[],
@@ -11,6 +13,7 @@ export type R2ObjectJSON = Omit<
   uploaded: string;
   checksums: R2StringChecksums;
   httpMetadata: Omit<R2HTTPMetadata, "cacheExpiry"> & {
+    // Original is typeof `Date`
     cacheExpiry?: string;
   };
 };
