+++ worker/vectorize/index.js
